🚚 rename: admin 폴더구조 변경 #63
