+++ src/pages/admin/DetailModal/components/DetailInfo.tsx
