+++ src/pages/admin/Registration/hooks/useRegistrationForm.ts
@@ -3,7 +3,7 @@ import { useForm } from "react-hook-form";
 import { zodResolver } from "@hookform/resolvers/zod";
 import useStore from "@/common/stores/useStore";
 import { FormValueType } from "@/pages/admin/type";
-import { regiSchema } from "@/pages/admin/validation/RegiSchema";
+import { regiSchema } from "@/pages/admin/Registration/validation/RegiSchema";
 import { getTeamId } from "@/common/utils/getTeamId";
 import alertToast from "@/common/utils/alertToast";
 import { MainMatchType } from "@/common/types/matchTypes";
